Modified the type heterogeneity check for set literals.  Now, it allows a mix of ints and floats rather than producing an error about nonhomogenous sets.
Also fixed a bug due to the behavior of Python sets: hash(1) and
hash(True) are both 1, so they can't coexist within the same set!
This subverted the old homogeneity checking code.

Also added some tests of sets with mixed ints and floats, and modified
the test related to homogeneity to ensure the {1, True} bug is
fixed.