MatchListener.matched() now returns the found bindings, instead of True/False.  Added MatchListener.get_sdos_from_binding() to convert a binding to a list of SDOs corresponding to the bound observations.
The match() function now returns SDOs contributing to a complete
match, if any.  There may be many ways of matching a pattern; it
returns the SDOs corresponding to the first binding in the list
returned by MatchListener.matched().

Added a few tests for checking that the right SDOs are returned.