perf: smarter stop condition for REPEATS
Previously [...] REPEATS k would not try ot generate a combination
if [...] produced less than k valuations.
This commit introduce a heuristics that bound the largest disjoint
combination that can be produced from [...] valuations.
REPEATS can stop if that number is smaller than k.
Example:
Consider [...] REPEATS 3
[...] is matched by (0,1), (1,2), (2,3), (3,0)
We have 4 valuations for [...], bigger than 3.
Yet it is clear that the size of the biggest disjoint combination is 2:
We have four values (0, 1, 2, 3) and each tuple consumes 2 of them.